+++ README.md
@@ -1,22 +1,45 @@
-# Next.js 12 React Server Components Demo (Alpha)
+# Management App:
 
-This is the demo of Hacker News built with Next.js and React Server Components. Read our announcement here: [Next.js 12](https://nextjs.org/blog/next-12).
+## Project Structure
 
-**Try the demo: https://next-news-rsc.vercel.sh**
+- Project Structure is inspired by this [post](https://www.robinwieruch.de/react-folder-structure/). Also follows the redux toolkit guideline for redux related code.
 
-### Development
+* `./pages/*`: Contains the pages of the Next App managed by the next router.
 
-To get started, run the following commands:
+* `./pages/api/*`: Contains the server side of the app that communicates with APIs and the Database
+
+* `./src/components/*`: Contains generic utility components like Button, Card, etc. They are meant to be re-used in other components and abstract common tasks.
+
+* `./src/features/*`: Contains components that have more complex logic and are features. Such components will use generic components from `./components/*`
+
+* `./server/*`: Contains server side code used in the `pages/api/*` files.
+
+* `./utils/*`: Contains helper functions.
+
+* `./data/*`: Contains data that is used across different files. Coule be named constants as well.
+
+* `./styled/*`: Contains styled components that can be re-used in the app.
+
+* `./typings/*`: Contains general typings that used across different files.
+
+* `./app/*`: Contains the Redux Store and typed `useDispatch` and `useSelector` hooks.
+
+* `./features/*`: Contains the redux reducers named `{{name}}.slice.ts`.
+
+## Note:
+
+> Furthermore, if a service/utils from the previous section is tightly coupled to a feature component, then move the service to the specific feature folder. The same may apply to other folders which were previously separated by technical concern:
+
+```jsx
+- src/
+--- feature/
+----- Payment/
+------- PaymentForm/
+------- PaymentWizard/
+------- services/
+--------- Currency/
+----------- index.js
+----------- service.js
+----------- test.js
 
 ```
-yarn
-yarn dev
-```
-
-And visit localhost:3000.
-
-### Note
-
-React Server Components are still [experimental](https://reactjs.org/blog/2020/12/21/data-fetching-with-react-server-components.html). To learn more about React Server Components, read our blog post: [Everything About React Server Components](https://vercel.com/blog/everything-about-react-server-components).
-
-React Server Components support is a built-in feature of Next.js 12. Full documentation is available here: [React 18 — Next.js](https://nextjs.org/docs/advanced-features/react-18).
